feat(flow): parse bridge errors into typed DescopeException
Move parseServerError from internal/http/ClientErrors.kt to
internal/others/Errors.kt — the parser is a general-purpose JSON-to-
DescopeException converter, not an HTTP concern. DescopeClient picks
it up at its new location; ClientErrors.kt is left with the HTTP
response-code handling that's actually its job.

Apply parseServerError to the JS-bridge error channel too:
FlowBridge.bridgeOnError now parses the incoming error string as a
typed DescopeException. Server-emitted flow cancellation (errorCode
E102122) collapses onto the existing DescopeException.flowCancelled
identity, carrying the server's errorMessage as the exception message
for context.

Relates to descope/etc#16441
