+++ descopesdk/src/main/java/com/descope/internal/http/ClientErrors.kt
@@ -1,21 +1,7 @@
 package com.descope.internal.http
 
-import com.descope.internal.others.toMap
 import com.descope.internal.others.with
 import com.descope.types.DescopeException
-import org.json.JSONObject
-
-internal fun parseServerError(response: String): DescopeException? {
-    try {
-        val map = JSONObject(response).toMap()
-        val code = map["errorCode"] as? String ?: return null
-        val desc = map["errorDescription"] as? String ?: "Descope server error"
-        val message = map["errorMessage"] as? String
-        return DescopeException(code = code, desc = desc, message = message)
-    } catch (_: Exception) {
-        return null
-    }
-}
 
 internal fun exceptionFromResponseCode(code: Int): DescopeException? {
     val desc = failureFromResponseCode(code) ?: return null
